Fix Git references at runtime.
Dothtml runs in the directory of the user's dot files, but needs
to run in a bundler context in order to avoid using the wrong gems.
To do so, it uses the Gemfile that is baked into the gem, however,
the Gemfile references the gemspec which in turn shells out to git.
A released gem does not have a .git directory, and this gives
warnings to STDERR.

The fix here is to avoid shelling out to git when the .git
directory doesn't exist.  This should not be a problem for
developers nor for deploying new versions.